+++ src/main/java/com/cyberoxi/hstpfacilities/models/responses/UnitReport.java
@@ -14,15 +14,15 @@ import java.util.List;
 @AllArgsConstructor
 public class UnitReport {
 
-    private int installmentsNumber; //تعداد اقساط وام
-    private int loanAmountPerMonth; //مبلغ وام در هر ماه
-    private int debtInstallmentsArrears; //مبلغ بدهی اقساط وام معوقه
+    private long installmentsNumber; //تعداد اقساط وام
+    private long loanAmountPerMonth; //مبلغ وام در هر ماه
+    private long debtInstallmentsArrears; //مبلغ بدهی اقساط وام معوقه
     private List<DateNumber> debtEstablishmentEachYear; //مبلغ بدهی استقرار به تفکیک سال
     private List<DateNumber> debtEstablishmentEachMonth; //مبلغ بدهی استقرار به تفکیک ماه
     private List<DateNumber> facilityPaidEachMonth; //مبلغ پرداختی تسهیلات به تفکیک هر ماه
     private List<DateNumber> establishmentPaidEachMonth; //مبلغ پرداختی استقرار به تفکیک هر ماه
-    private int totalFacilityAmountPaid; //کل مبلغ پرداختی تسهیلات
-    private int totalEstablishmentAmountPaid; //کل مبلغ پرداختی استقرار
-    private int debtFacilityRemained; //مانده بدهی تسهیلات
-    private int debtEstablishmentRemained; //مانده بدهی استقرار
+    private long totalFacilityAmountPaid; //کل مبلغ پرداختی تسهیلات
+    private long totalEstablishmentAmountPaid; //کل مبلغ پرداختی استقرار
+    private long debtFacilityRemained; //مانده بدهی تسهیلات
+    private long debtEstablishmentRemained; //مانده بدهی استقرار
 }
